Move UFO merger to library so it can be used in other scripts
